Fix bash ignorance
It is an utterly ignorant myth that bash is installed on every
Unix system and always called /bin/bash. I have no problems with
bash as a dependency, but since it is just a user application
as any other user application, there is absolutely no reason why
it should not be installed in /usr/local/bin, as are most user
applications.

The only shell where you may hardcode a path to is the Posix
shell called /bin/sh (and don't assume that it is a bash).