List of dependencies.
This could be the start of a general instruction, on how to build and install the
software. Please contribute. Also note that the list might be incomplete, but all
software mentioned is needed (not necessarily in the versions mentioned). So this
can be a good starting point of a full README file.