Working on translation of equations in C code from Python.
I have to do this as the OCaml parser is rather limited and fails when
encountering divisions, etc. As the Python version is already better,
I'll focus on improving it further. It will generate the C code by itself.
At some point I may even work so that the Jacobian is computed from Python.
The aim is to finally shut down the OCaml parser.
This is only used in Nmag5 (nsim.model).

As a first step: special tensors with numerical indices can now be evaluated
on the fly (and simplified).